LPS-194013 Inline DDMStructureIndexerRegistry into DDMStructureLocalServiceImpl, register DDMStructureIndexer in the message and get it in DDMStructureReindexMessageListener and fix tests
