+++ modules/apps/document-library/document-library-test/src/testIntegration/java/com/liferay/document/library/search/test/DLFileEntryMetadataDDMStructureIndexerTest.java
@@ -13,6 +13,7 @@ import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.messaging.MessageBus;
+import com.liferay.portal.kernel.search.DDMStructureIndexer;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
@@ -92,6 +93,8 @@ public class DLFileEntryMetadataDDMStructureIndexerTest
 
 		message.put("structureId", ddmStructure.getStructureId());
 
+		message.put("ddmStructureIndexer", _ddmStructureIndexer);
+
 		_messageBus.sendMessage("liferay/ddm_structure_reindex", message);
 
 		dlSearchFixture.searchOnlyOne(searchTerm, LocaleUtil.JAPAN);
@@ -124,5 +127,10 @@ public class DLFileEntryMetadataDDMStructureIndexerTest
 
 	@Inject
 	private static MessageBus _messageBus;
+
+	@Inject(
+		filter = "ddm.structure.indexer.class.name=com.liferay.document.library.kernel.model.DLFileEntryMetadata"
+	)
+	private DDMStructureIndexer _ddmStructureIndexer;
 
 }
