+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/messaging/DDMStructureReindexMessageListener.java
@@ -5,7 +5,6 @@
 
 package com.liferay.dynamic.data.mapping.internal.messaging;
 
-import com.liferay.dynamic.data.mapping.internal.background.task.DDMStructureIndexerRegistry;
 import com.liferay.dynamic.data.mapping.internal.constants.DDMDestinationNames;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
@@ -33,12 +32,8 @@ public class DDMStructureReindexMessageListener extends BaseMessageListener {
 	protected void doReceive(Message message) throws Exception {
 		long structureId = message.getLong("structureId");
 
-		DDMStructure structure = _ddmStructureLocalService.getStructure(
-			structureId);
-
-		DDMStructureIndexer structureIndexer =
-			_ddmStructureIndexerRegistry.getDDMStructureIndexer(
-				structure.getClassName());
+		DDMStructureIndexer structureIndexer = (DDMStructureIndexer)message.get(
+			"ddmStructureIndexer");
 
 		List<Long> ddmStructureIds = new ArrayList<>();
 
@@ -62,9 +57,6 @@ public class DDMStructureReindexMessageListener extends BaseMessageListener {
 				structureIds, structure.getStructureId());
 		}
 	}
-
-	@Reference
-	private DDMStructureIndexerRegistry _ddmStructureIndexerRegistry;
 
 	@Reference
 	private DDMStructureLocalService _ddmStructureLocalService;
